favorites

+++ app/src/main/java/com/example/saschag/lerningkotlin/ArticleActivity.kt
@@ -10,7 +10,7 @@ class ArticleActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_article)
 
-        val id = intent.getLongExtra("article_id", 0)
+        val id = intent.getIntExtra("article_id", 0)
         val cursor = DBHelper(applicationContext).getArticleById(id)
         if (cursor != null && cursor.moveToFirst()) {
             findViewById<TextView>(R.id.textViewTitle).text = cursor.getString(cursor.getColumnIndex(DBTables.Articles.COLUMN_TITLE))
